requirements.txt
python-pptx>=0.6.21,<2
python-docx>=1.1,<2
pypandoc>=1.13,<2
Flask>=2.3,<4
python-dotenv>=1.0,<2
Pillow>=10.0,<12
openai>=1.0,<2
